Rename revokePermission function to apmRevokePermission
