+++ packages/toolkit/src/apm/index.js
@@ -6,5 +6,5 @@ export { default as grantNewVersionsPermission } from './grantNewVersionsPermiss
 export { default as apmPublishVersion } from './apmPublishVersion'
 export { default as apmPublishVersionIntent } from './apmPublishVersionIntent'
 export { default as apmGetFile } from './apmGetFile'
-export { default as revokePermission } from './revokePermission'
+export { default as apmRevokePermission } from './apmRevokePermission'
 export { default as apmSetPermissionManager } from './apmSetPermissionManager'
